+++ x/claims/keeper/claim_test.go
@@ -187,15 +187,14 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 	addr := sdk.AccAddress(tests.GenerateAddress().Bytes())
 
 	testCases := []struct {
-		name            string
-		malleate        func()
-		claimsRecord    types.ClaimsRecord
-		action          types.Action
-		params          types.Params
-		expAmt          sdk.Int
-		expClawedBack   sdk.Int
-		expError        bool
-		expDeleteRecord bool
+		name          string
+		malleate      func()
+		claimsRecord  types.ClaimsRecord
+		action        types.Action
+		params        types.Params
+		expAmt        sdk.Int
+		expClawedBack sdk.Int
+		expError      bool
 	}{
 		{
 			"fail - unspecified action",
@@ -206,7 +205,6 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 			sdk.ZeroInt(),
 			sdk.ZeroInt(),
 			true,
-			false,
 		},
 		{
 			"zero - claims inactive",
@@ -216,7 +214,6 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 			types.Params{},
 			sdk.ZeroInt(),
 			sdk.ZeroInt(),
-			false,
 			false,
 		},
 		{
@@ -234,7 +231,6 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 			sdk.ZeroInt(),
 			sdk.ZeroInt(),
 			false,
-			false,
 		},
 		{
 			"zero - claimable amount is 0",
@@ -250,7 +246,6 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 			},
 			sdk.ZeroInt(),
 			sdk.ZeroInt(),
-			false,
 			false,
 		},
 		{
@@ -274,7 +269,6 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 			sdk.ZeroInt(),
 			sdk.ZeroInt(),
 			true,
-			false,
 		},
 		{
 			"failed - error during community pool fund",
@@ -297,7 +291,6 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 			sdk.ZeroInt(),
 			sdk.ZeroInt(),
 			true,
-			false,
 		},
 		{
 			"success - claim single action",
@@ -320,7 +313,6 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 			sdk.NewInt(50),
 			sdk.ZeroInt(),
 			false,
-			false,
 		},
 		{
 			"success - claim single action during decay",
@@ -342,7 +334,6 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 			},
 			sdk.NewInt(25),
 			sdk.NewInt(25),
-			false,
 			false,
 		},
 		{
@@ -369,7 +360,6 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 			sdk.NewInt(50),
 			sdk.ZeroInt(),
 			false,
-			true,
 		},
 	}
 
@@ -405,13 +395,9 @@ func (suite *KeeperTestSuite) TestClaimCoinsForAction() {
 				suite.Require().Equal(initialCommunityPoolCoins.String(), funds.String())
 			}
 
-			if tc.expDeleteRecord {
-				suite.Require().False(suite.app.ClaimsKeeper.HasClaimsRecord(suite.ctx, addr))
-			} else {
-				cr, found := suite.app.ClaimsKeeper.GetClaimsRecord(suite.ctx, addr)
-				suite.Require().True(found)
-				suite.Require().True(cr.HasClaimedAction(tc.action))
-			}
+			cr, found := suite.app.ClaimsKeeper.GetClaimsRecord(suite.ctx, addr)
+			suite.Require().True(found)
+			suite.Require().True(cr.HasClaimedAction(tc.action))
 		})
 	}
 }
@@ -502,37 +488,6 @@ func (suite *KeeperTestSuite) TestMergeClaimRecords() {
 			},
 		},
 		{
-			"case 3: sender completed action",
-			func() {
-				senderClaimsRecord := types.NewClaimsRecord(sdk.NewInt(200))
-				senderClaimsRecord.MarkClaimed(types.ActionVote)
-				recipientClaimsRecord := types.NewClaimsRecord(sdk.NewInt(200))
-
-				expBalance := suite.app.BankKeeper.GetBalance(suite.ctx, recipient, params.ClaimsDenom)
-
-				coins := sdk.Coins{sdk.NewCoin(params.ClaimsDenom, sdk.NewInt(150))}
-				err := suite.app.BankKeeper.MintCoins(suite.ctx, inflationtypes.ModuleName, coins)
-				suite.Require().NoError(err)
-				err = suite.app.BankKeeper.SendCoinsFromModuleToModule(suite.ctx, inflationtypes.ModuleName, types.ModuleName, coins)
-				suite.Require().NoError(err)
-
-				mergedRecord, err := suite.app.ClaimsKeeper.MergeClaimsRecords(suite.ctx, recipient, senderClaimsRecord, recipientClaimsRecord, params)
-				suite.Require().NoError(err)
-
-				// Vote and IBC actions should be claimed
-				expectedRecord := types.NewClaimsRecord(sdk.NewInt(400))
-				expectedRecord.MarkClaimed(types.ActionVote)
-				expectedRecord.MarkClaimed(types.ActionIBCTransfer)
-
-				suite.Require().Equal(expectedRecord, mergedRecord)
-
-				expBalance = expBalance.Add(sdk.NewCoin(params.ClaimsDenom, sdk.NewInt(150)))
-
-				balance := suite.app.BankKeeper.GetBalance(suite.ctx, recipient, params.ClaimsDenom)
-				suite.Require().Equal(expBalance, balance)
-			},
-		},
-		{
 			"case 2: recipient completed all actions, but IBC transfer",
 			func() {
 				senderClaimsRecord := types.ClaimsRecord{
@@ -569,7 +524,7 @@ func (suite *KeeperTestSuite) TestMergeClaimRecords() {
 			},
 		},
 		{
-			"case 1: sender and recipient completed all",
+			"fail - sender and recipient completed all",
 			func() {
 				senderClaimsRecord := types.ClaimsRecord{
 					InitialClaimableAmount: sdk.NewInt(200),
@@ -580,20 +535,8 @@ func (suite *KeeperTestSuite) TestMergeClaimRecords() {
 					ActionsCompleted:       []bool{true, true, true, true},
 				}
 
-				expBalance := suite.app.BankKeeper.GetBalance(suite.ctx, recipient, params.ClaimsDenom)
-
-				mergedRecord, err := suite.app.ClaimsKeeper.MergeClaimsRecords(suite.ctx, recipient, senderClaimsRecord, recipientClaimsRecord, params)
-				suite.Require().NoError(err)
-
-				expectedRecord := types.ClaimsRecord{
-					InitialClaimableAmount: sdk.NewInt(400),
-					ActionsCompleted:       []bool{true, true, true, true},
-				}
-
-				suite.Require().Equal(expectedRecord, mergedRecord)
-
-				balance := suite.app.BankKeeper.GetBalance(suite.ctx, recipient, params.ClaimsDenom)
-				suite.Require().Equal(expBalance, balance)
+				_, err := suite.app.ClaimsKeeper.MergeClaimsRecords(suite.ctx, recipient, senderClaimsRecord, recipientClaimsRecord, params)
+				suite.Require().Error(err)
 			},
 		},
 		{